Kuhn's algo based on e-maxx
It's assumed that an input graph is already splitted into two parts, and edges from left to right are stored explicitly.
An algorithm is an application of Berge's theorem: a matching is maximal iff there is no augmenting (both starting and ending with an edge not from a matching such that every adjacent pair of edges on the path contains one edge from the matching and one edge outside of it). To improve the efficiency of computations, store an array (mt[] from "matching") that shows for every node from the right part whether it is already a part of a matching and an id of the corresponding node in the left part in that case.
To find such path, use an efficient dfs-based approach: start with some rough approximate matching and then consequently try every node from the left part: if is it not contained in the matching, try to enhance the matching starting by finding some augmenting path (recursively) with a beginning in this node: look over all reachable nodes and if any of them is not in a matching, you are done; otherwise, go by an edge from the matching to the node from the left part that is connected to current, and repeat the procedure once again from that node.